requirements.txt
youtool==0.2.0
mongomock==4.0.0
pymongo==3.12.1
pytest==7.2.2
pytest-mock==3.10.0
python-dotenv>=0.21.0
yt-dlp>=2023.10.13
webvtt-py>=0.4.6
black>=23.3.0
flake8>=6.0.0
